embed_server/requirements.txt
flask
gunicorn
transformers
accelerate
Pillow
commentjson
numpy
